Update main.js
Changes made appear to fix errors caused by switching from an accepted HTML file type to a different file type and then back to an HTML file. Brackets appears to run KeyBindingManager commands when adding menu items with keyboard shortcuts but does not do the same when removing menu items, so it must be done manually.