Quick fix for older Brackets versions
Recently noticed that Brackets 1.7 has a deprecation warning for Mustache being used in the JSLint directives and that it should be loaded with the brackets.getModule() method. Made that change but came into work Monday morning with it breaking older versions of Brackets. Reverting back for now until I figure out the greater depth of Brackets extensions and if they allow versioning based on Brackets app version or not.